requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
Django==2.1.1
idna==2.9
mysql-connector==2.2.9
pytz==2020.1
requests==2.23.0
urllib3==1.25.9
